# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SWAP API Documantation",
        "version": "latest"
    },
    "paths": {
...
        "/api/swap": {
            "get": {
                "tags": [
                    "Swaps"
                ],
                "summary": "Swap List.",
                "operationId": "553ef66146bdc1744b27c08d985c38f5",
                "parameters": [
                    {
                        "name": "pagination",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "10"
                    },
                    {
                        "name": "get_all",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "json",
                                            "example": []
                                        },
                                        "links": {
                                            "type": "json",
                                            "example": []
                                        },
                                        "meta": {
                                            "type": "json",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "errors": {
                                            "type": "json",
                                            "example": {
                                                "message": [
                                                    "Unauthenticated"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Swaps"
                ],
                "summary": "Create a new swap",
                "description": "Create Swap.",
                "operationId": "82590266f8107084d1a4b8f21149c9b2",
                "parameters": [
                    {
                        "name": "requested_user_id",
                        "in": "query",
                        "description": "Requested User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "exchanged_user_id",
                        "in": "query",
                        "description": "Exchanged User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "pending"
                        }
                    },
                    {
                        "name": "requested_wholesale_amount",
                        "in": "query",
                        "description": "Requested Wholesale Amount",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 100
                        }
                    },
                    {
                        "name": "exchanged_wholesale_amount",
                        "in": "query",
                        "description": "Exchanged Wholesale Amount",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 120
                        }
                    },
                    {
                        "name": "requested_total_commission",
                        "in": "query",
                        "description": "Requested Total Commission",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double",
                            "example": 10.5
                        }
                    },
                    {
                        "name": "exchanged_total_commission",
                        "in": "query",
                        "description": "Exchanged Total Commission",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double",
                            "example": 12.5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "errors": {
                                            "type": "json",
                                            "example": {
                                                "message": [
                                                    "Swap created successfully."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "errors": {
                                            "type": "json",
                                            "example": {
                                                "message": [
                                                    "The given data was invalid."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    }
}